Rename `crate::commands` to `crate::command`
