+++ src/lib.rs
@@ -2,7 +2,7 @@
 // License, v. 2.0. If a copy of the MPL was not distributed with this
 // file, You can obtain one at https://mozilla.org/MPL/2.0/.
 
-pub mod commands;
+pub mod command;
 pub mod options;
 
 pub(crate) mod analyzer;
